# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a proxy server LOL API. You can find the repo fo this project (https://github.com/emreunsal1/lol-backend).<br><br> Contact: https://www.linkedin.com/in/emreunsal111/",
    "version": "1.0.0",
    "title": "LoL API",
    "contact": {
      "email": "unsalemre111@gmail.com"
    }
  },
  "host": "leagueoflegends-backend.herokuapp.com",
  "basePath": "/api",
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/match/detail/{matchid}": {
      "get": {
        "tags": [
          "Match"
        ],
        "summary": "Log User Match Detail",
        "description": "",
        "operationId": "matchDetail",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "matchid",
            "in": "path",
            "description": "The match id for match detail ",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid username/password supplied"
          }
        }
      }
    },
...
  }
}